use seperate column for actions
