+++ packages/form-config-antd/src/components/DraftFileList/utils.tsx
@@ -1,26 +1,13 @@
-import React from 'react';
 import { Dictionary } from '@onaio/utils';
-import { ManifestFilesTypes, formatDate } from '@opensrp/form-config-core';
-import { TableActions } from './TableActions';
-import { getFetchOptions } from '@opensrp/server-service';
+import { formatDate } from '@opensrp/form-config-core';
 
 /**
  * Return table columns
  *
- * @param {string} accessToken Opensrp API access token
- * @param {string} opensrpBaseURL Opensrp API base URL
- * @param {boolean} isJsonValidator true if is json validator, false otherwise
  * @param {Dictionary} sortedInfo object containing sort order information
- * @param {Function} customFetchOptions Opensrp API custom options
  * @returns {Dictionary[]} table columns
  */
-export const getTableColumns = (
-  accessToken: string,
-  opensrpBaseURL: string,
-  isJsonValidator: boolean,
-  sortedInfo?: Dictionary,
-  customFetchOptions?: typeof getFetchOptions
-): Dictionary[] => {
+export const getTableColumns = (sortedInfo?: Dictionary): Dictionary[] => {
   const columns: Dictionary[] = [];
   const headerItems: string[] = ['Identifier', 'File Name', 'File Version', 'Created At', 'Module'];
   const fields: string[] = ['identifier', 'label', 'version', 'createdAt', 'module'];
@@ -47,20 +34,5 @@ export const getTableColumns = (
     columns.push(column);
   });
 
-  columns.push({
-    title: 'Action',
-    key: 'action',
-    // eslint-disable-next-line react/display-name
-    render: (_: string, file: ManifestFilesTypes) => {
-      const tableActionProps = {
-        file,
-        accessToken,
-        opensrpBaseURL,
-        isJsonValidator,
-        customFetchOptions,
-      };
-      return <TableActions {...tableActionProps} />;
-    },
-  });
   return columns;
 };
