+++ packages/form-config-antd/src/components/FileList/utils.tsx
@@ -1,29 +1,18 @@
-import React from 'react';
 import { Dictionary } from '@onaio/utils';
-import { ManifestFilesTypes, formatDate } from '@opensrp/form-config-core';
-import { TableActions } from './TableActions';
-import { getFetchOptions } from '@opensrp/server-service';
+import { formatDate } from '@opensrp/form-config-core';
 import lang, { Lang } from '../../lang';
 
 /**
  * Return table columns
  *
- * @param accessToken  Opensrp API access token
- * @param opensrpBaseURL Opensrp API base URL
  * @param isJsonValidator boolean to check whether is Json validator
- * @param uploadFileURL route to upload form
  * @param sortedInfo object containing sort order information
- * @param customFetchOptions custom Opensrp API fetch options
  * @param langObj the language translation object
  * @returns {Dictionary[]} table columns
  */
 export const getTableColumns = (
-  accessToken: string,
-  opensrpBaseURL: string,
   isJsonValidator: boolean,
-  uploadFileURL: string,
   sortedInfo?: Dictionary,
-  customFetchOptions?: typeof getFetchOptions,
   langObj: Lang = lang
 ): Dictionary[] => {
   const columns: Dictionary[] = [];
@@ -65,21 +54,5 @@ export const getTableColumns = (
     });
   }
 
-  columns.push({
-    title: langObj.ACTION,
-    key: 'action',
-    // eslint-disable-next-line react/display-name
-    render: (_: string, file: ManifestFilesTypes) => {
-      const tableActionProps = {
-        file,
-        uploadFileURL,
-        accessToken,
-        opensrpBaseURL,
-        isJsonValidator,
-        customFetchOptions,
-      };
-      return <TableActions {...tableActionProps} />;
-    },
-  });
   return columns;
 };
