+++ packages/react-utils/src/components/TableLayout/index.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { Table as AntTable } from 'antd';
 import { ColumnType, TableProps as AntTableProps } from 'antd/lib/table';
 import { Dictionary } from '@onaio/utils';
-import { TABLE_PAGE_SIZE, TABLE_PAGE_SIZE_OPTIONS } from '../../constants';
+import { TABLE_PAGE_SIZE, TABLE_PAGE_SIZE_OPTIONS, TABLE_ACTIONS_KEY } from '../../constants';
 import { getConfig, TableState, setConfig } from '@opensrp/pkg-config';
 
 type TKey<T> = keyof T & React.Key;
@@ -28,7 +28,10 @@ export const defaults: Options = {
 interface Props<T> extends Omit<Options<T>, 'columns' | 'dataSource'> {
   datasource: T[];
   columns?: Column<T>[];
+  actions?: Optional<Column<T>, 'key' | 'dataIndex'>;
 }
+
+type Optional<T, K extends keyof T> = Omit<T, K> & Partial<T>;
 
 interface PersistState {
   id: string;
@@ -48,10 +51,19 @@ export type TableProps<T> = Props<T> & (PersistState | NoPersistState);
  * @returns - the component
  */
 export function TableLayout<T extends object = Dictionary>(props: TableProps<T>) {
-  const { id, columns, datasource, children, persistState, ...restprops } = props;
+  const { id, columns, datasource, children, persistState, actions, ...restprops } = props;
 
   const options: Options = { ...defaults, ...restprops };
   const tablesState = getConfig('tablespref') ?? {};
+
+  if (columns && actions) {
+    const actionsColumn: Column<T> = {
+      key: TABLE_ACTIONS_KEY as TKey<T>,
+      title: 'Actions',
+      ...actions,
+    };
+    columns.push(actionsColumn);
+  }
 
   const [tableState, setTableState] = useState<TableState>(
     id && tablesState[id] !== undefined ? tablesState[id] : {}
