+++ packages/react-utils/src/constants.ts
@@ -1,5 +1,7 @@
 export const HOME_URL = '/';
 export const EXPRESS_TOKEN_REFRESH_URL = '/refresh/token';
 
+export const TABLE_ACTIONS_KEY = '_actions';
+
 export const TABLE_PAGE_SIZE = 5;
 export const TABLE_PAGE_SIZE_OPTIONS = ['5', '10', '20', '50', '100'];
